LPS-181640 Rewrite Header to use Atoms
